Fix stream loading reporting
The symbols used for "kilo-" were incorrect, so this commit fixes
them. Also, since the SI prefixes are being used (rightly so for
any user-facing text), the bytes should be divided accordingly by
powers of 10 rather than powers of 2, since data can come in
non-powers of 2. It moves the associated strings out of the
application strings file because they are SI symbols, which are
the same in all languages, so they don't need localization.

The commit also moves the formatting for the number of seeds into
the application strings file for better localization.